+++ src/NuGet.Core/NuGet.CommandLine.XPlat/Commands/WhyCommand/WhyCommandRunner.cs
@@ -9,9 +9,8 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using Microsoft.Build.Evaluation;
-using NuGet.Frameworks;
+using NuGet.CommandLine.XPlat.WhyCommandUtility;
 using NuGet.ProjectModel;
-using NuGet.Shared;
 
 namespace NuGet.CommandLine.XPlat
 {
@@ -30,15 +29,12 @@ namespace NuGet.CommandLine.XPlat
             {
                 ValidatePathArgument(whyCommandArgs.Path);
                 ValidatePackageArgument(whyCommandArgs.Package);
-                ValidateFrameworksOption(whyCommandArgs.Frameworks);
             }
             catch (ArgumentException ex)
             {
                 whyCommandArgs.Logger.LogError(ex.Message);
                 return ExitCodes.InvalidArguments;
             }
-
-            var msBuild = new MSBuildAPIUtility(whyCommandArgs.Logger);
 
             string targetPackage = whyCommandArgs.Package;
 
@@ -53,7 +49,26 @@ namespace NuGet.CommandLine.XPlat
 
                 if (assetsFile != null)
                 {
-                    OutputAllDependencyGraphs(whyCommandArgs, msBuild, project, assetsFile);
+                    Dictionary<string, List<DependencyNode>?>? dependencyGraphPerFramework = DependencyGraphFinder.GetAllDependencyGraphs(whyCommandArgs, assetsFile, project.DirectoryPath);
+
+                    if (dependencyGraphPerFramework != null)
+                    {
+                        whyCommandArgs.Logger.LogMinimal(
+                            string.Format(
+                                Strings.WhyCommand_Message_DependencyGraphsFoundInProject,
+                                project.GetPropertyValue(ProjectName),
+                                targetPackage));
+
+                        DependencyGraphPrinter.PrintAllDependencyGraphs(dependencyGraphPerFramework, targetPackage, whyCommandArgs.Logger);
+                    }
+                    else
+                    {
+                        whyCommandArgs.Logger.LogMinimal(
+                            string.Format(
+                                Strings.WhyCommand_Message_NoDependencyGraphsFoundInProject,
+                                project.GetPropertyValue(ProjectName),
+                                targetPackage));
+                    }
                 }
 
                 ProjectCollection.GlobalProjectCollection.UnloadProject(project);
@@ -92,39 +107,6 @@ namespace NuGet.CommandLine.XPlat
                     Strings.WhyCommand_Error_ArgumentCannotBeEmpty,
                     "PACKAGE"));
             }
-        }
-
-        private static void ValidateFrameworksOption(List<string> frameworks)
-        {
-            var parsedFrameworks = frameworks.Select(f =>
-                                    NuGetFramework.Parse(
-                                        f.Split(new char[] { '/' }, StringSplitOptions.RemoveEmptyEntries)
-                                            .Select(s => s.Trim())
-                                            .ToArray()[0]));
-
-            if (parsedFrameworks.Any(f => f.Framework.Equals("Unsupported", StringComparison.OrdinalIgnoreCase)))
-            {
-                throw new ArgumentException(
-                    string.Format(CultureInfo.CurrentCulture,
-                    Strings.WhyCommand_Error_InvalidFramework));
-            }
-
-            // Break up the whole thing into a for loop instead of using linq, and then we can output the errors for each user input framework,
-            // OR
-            // Remove this validation completely, and get targets based on whatever inputs the user provides.
-            // Still need to validate that the user input 'target alias' exists in the assets file
-            /*
-            foreach (var framework in parsedFrameworks)
-            {
-                if (framework.Framework.Equals("Unsupported", StringComparison.OrdinalIgnoreCase))
-                {
-                    throw new ArgumentException(
-                    string.Format(CultureInfo.CurrentCulture,
-                    Strings.WhyCommand_Error_InvalidFramework,
-                    framework.GetShortFolderName()));
-                }
-            }
-            */
         }
 
         /// <summary>
@@ -178,324 +160,6 @@ namespace NuGet.CommandLine.XPlat
             }
 
             return assetsFile;
-        }
-
-        /// <summary>
-        /// Finds and prints dependency graphs for a given project.
-        /// </summary>
-        /// <param name="whyCommandArgs">CLI arguments for the 'why' command</param>
-        /// <param name="msBuild">MSBuild utility</param>
-        /// <param name="project">Current project</param>
-        /// <param name="assetsFile">Assets file for current project</param>
-        private static void OutputAllDependencyGraphs(
-            WhyCommandArgs whyCommandArgs,
-            MSBuildAPIUtility msBuild,
-            Project project,
-            LockFile assetsFile)
-        {
-            bool doesProjectHaveDependencyOnPackage = false;
-            string targetPackage = whyCommandArgs.Package;
-            var dependencyGraphPerFramework = new Dictionary<string, List<DependencyNode>?>(assetsFile.Targets.Count);
-
-            // get all resolved package references for a project
-            List<FrameworkPackages> frameworkPackages = msBuild.GetResolvedVersions(project, whyCommandArgs.Frameworks, assetsFile, transitive: true, includeProjectReferences: true);
-            Dictionary<string, List<string>> topLevelReferences = GetTopLevelPackageAndProjectReferences(assetsFile, whyCommandArgs.Frameworks, project.DirectoryPath);
-
-            if (frameworkPackages?.Count > 0)
-            {
-                foreach (var frameworkPackage in frameworkPackages)
-                {
-                    LockFileTarget target = assetsFile.GetTarget(frameworkPackage.Framework, runtimeIdentifier: null);
-
-                    if (target != default)
-                    {
-                        // get all package libraries for the framework
-                        IList<LockFileTargetLibrary> packageLibraries = target.Libraries;
-
-                        // if the project has a dependency on the target package, get the dependency graph
-                        if (packageLibraries.Any(l => l.Name == targetPackage))
-                        {
-                            doesProjectHaveDependencyOnPackage = true;
-                            dependencyGraphPerFramework.Add(frameworkPackage.Framework,
-                                                            GetDependencyGraphPerFramework(frameworkPackage.TopLevelPackages, packageLibraries, targetPackage));
-                        }
-                        else
-                        {
-                            dependencyGraphPerFramework.Add(frameworkPackage.Framework, null);
-                        }
-                    }
-                }
-            }
-
-            if (!doesProjectHaveDependencyOnPackage)
-            {
-                whyCommandArgs.Logger.LogMinimal(
-                    string.Format(
-                        Strings.WhyCommand_Message_NoDependencyGraphsFoundInProject,
-                        project.GetPropertyValue(ProjectName),
-                        targetPackage));
-            }
-            else
-            {
-                whyCommandArgs.Logger.LogMinimal(
-                    string.Format(
-                        Strings.WhyCommand_Message_DependencyGraphsFoundInProject,
-                        project.GetPropertyValue(ProjectName),
-                        targetPackage));
-
-                WhyCommandPrintUtility.PrintAllDependencyGraphs(dependencyGraphPerFramework, targetPackage, whyCommandArgs.Logger);
-            }
-        }
-
-        /// <summary>
-        /// Finds all dependency paths from the top-level packages to the target package for a given framework.
-        /// </summary>
-        /// <param name="topLevelPackages">All top-level packages for the framework.</param>
-        /// <param name="packageLibraries">All package libraries for a given framework.</param>
-        /// <param name="targetPackage">The package we want the dependency paths for.</param>
-        /// <returns>List of all top-level package nodes in the dependency graph.</returns>
-        private static List<DependencyNode>? GetDependencyGraphPerFramework(
-            IEnumerable<InstalledPackageReference> topLevelPackages,
-            IList<LockFileTargetLibrary> packageLibraries,
-            string targetPackage)
-        {
-            List<DependencyNode>? dependencyGraph = null;
-
-            // hashset tracking every package node that we've traversed
-            var visited = new HashSet<string>();
-            // dictionary tracking all package nodes that have been added to the graph, mapped to their DependencyNode objects
-            var dependencyNodes = new Dictionary<string, DependencyNode>();
-            // dictionary mapping all packageIds to their resolved version
-            Dictionary<string, string> versions = GetAllResolvedVersions(packageLibraries);
-
-            foreach (var topLevelPackage in topLevelPackages)
-            {
-                // use depth-first search to find dependency paths from the top-level package to the target package
-                DependencyNode? topLevelNode = FindDependencyPath(topLevelPackage.Name, packageLibraries, visited, dependencyNodes, versions, targetPackage);
-
-                if (topLevelNode != null)
-                {
-                    dependencyGraph ??= [];
-                    dependencyGraph.Add(topLevelNode);
-                }
-            }
-
-            return dependencyGraph;
-        }
-
-
-        /// <summary>
-        /// Adds all resolved versions of packages to a dictionary.
-        /// </summary>
-        /// <param name="packageLibraries">All package libraries for a given framework.</param>
-        private static Dictionary<string, string> GetAllResolvedVersions(IList<LockFileTargetLibrary> packageLibraries)
-        {
-            var versions = new Dictionary<string, string>();
-
-            foreach (var package in packageLibraries)
-            {
-                if (package?.Name != null && package?.Version != null)
-                {
-                    versions.Add(package.Name, package.Version.ToNormalizedString());
-                }
-            }
-
-            return versions;
-        }
-
-        /// <summary>
-        /// Traverses the dependency graph for a given top-level package, looking for a path to the target package.
-        /// </summary>
-        /// <param name="topLevelPackage">Top-level package.</param>
-        /// <param name="packageLibraries">All package libraries for a given framework.</param>
-        /// <param name="visited">HashSet tracking every package node that we've traversed.</param>
-        /// <param name="dependencyNodes">Dictionary tracking all packageIds that were added to the graph, mapped to their DependencyNode objects.</param>
-        /// <param name="versions">Dictionary mapping packageIds to their resolved versions.</param>
-        /// <param name="targetPackage">The package we want the dependency paths for.</param>
-        /// <returns>The top-level package node in the dependency graph (if a path was found), or null (if no path was found)</returns>
-        private static DependencyNode? FindDependencyPath(
-            string topLevelPackage,
-            IList<LockFileTargetLibrary> packageLibraries,
-            HashSet<string> visited,
-            Dictionary<string, DependencyNode> dependencyNodes,
-            Dictionary<string, string> versions,
-            string targetPackage)
-        {
-            var stack = new Stack<StackDependencyData>();
-            stack.Push(new StackDependencyData(topLevelPackage, null));
-
-            while (stack.Count > 0)
-            {
-                var currentPackageData = stack.Pop();
-                var currentPackageId = currentPackageData.Id;
-
-                // if we reach the target node, or if we've already traversed this node and found dependency paths, add it to the graph
-                if (currentPackageId == targetPackage
-                    || dependencyNodes.ContainsKey(currentPackageId))
-                {
-                    AddToGraph(currentPackageData, dependencyNodes, versions);
-                    continue;
-                }
-
-                // if we have already traversed this node's children, continue
-                if (visited.Contains(currentPackageId))
-                {
-                    continue;
-                }
-
-                visited.Add(currentPackageId);
-
-                // get all dependencies for the current package
-                var dependencies = packageLibraries?.FirstOrDefault(i => i.Name == currentPackageId)?.Dependencies;
-
-                if (dependencies?.Count > 0)
-                {
-                    // push all the dependencies onto the stack
-                    foreach (var dependency in dependencies)
-                    {
-                        stack.Push(new StackDependencyData(dependency.Id, currentPackageData));
-                    }
-                }
-            }
-
-            if (dependencyNodes.ContainsKey(topLevelPackage))
-            {
-                return dependencyNodes[topLevelPackage];
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Adds a dependency path to the graph, starting from the target package and traversing up to the top-level package.
-        /// </summary>
-        /// <param name="targetPackageData">Target node data. This stores parent references, so it can be used to construct the dependency graph
-        /// up to the top-level package.</param>
-        /// <param name="dependencyNodes">Dictionary tracking all packageIds that were added to the graph, mapped to their DependencyNode objects.</param>
-        /// <param name="versions">Dictionary mapping packageIds to their resolved versions.</param>
-        private static void AddToGraph(
-            StackDependencyData targetPackageData,
-            Dictionary<string, DependencyNode> dependencyNodes,
-            Dictionary<string, string> versions)
-        {
-            // first, we traverse the target's parents, listing the packages in the path from the target to the top-level package
-            var dependencyPath = new List<string> { targetPackageData.Id };
-            StackDependencyData? current = targetPackageData.Parent;
-
-            while (current != null)
-            {
-                dependencyPath.Add(current.Id);
-                current = current.Parent;
-            }
-
-            // then, we traverse this list from the target package to the top-level package, initializing/updating their dependency nodes as needed
-            for (int i = 0; i < dependencyPath.Count; i++)
-            {
-                string currentPackageId = dependencyPath[i];
-
-                if (!dependencyNodes.ContainsKey(currentPackageId))
-                {
-                    dependencyNodes.Add(currentPackageId, new DependencyNode(currentPackageId, versions[currentPackageId]));
-                }
-
-                if (i > 0)
-                {
-                    var childNode = dependencyNodes[dependencyPath[i - 1]];
-
-                    if (dependencyNodes[currentPackageId].Children.Any(p => p.Id == childNode.Id))
-                    {
-                        continue;
-                    }
-
-                    dependencyNodes[currentPackageId].Children.Add(childNode);
-                }
-            }
-        }
-
-        private static Dictionary<string, List<string>> GetTopLevelPackageAndProjectReferences(LockFile assetsFile, List<string> userInputFrameworks, string projectDirectoryPath)
-        {
-            var topLevelReferences = new Dictionary<string, List<string>>();
-
-            var targetAliases = assetsFile.PackageSpec.RestoreMetadata.OriginalTargetFrameworks;
-
-            if (userInputFrameworks?.Count > 0)
-            {
-                targetAliases = targetAliases.Where(f => userInputFrameworks.Contains(f)).ToList();
-            }
-
-            var projectLibraries = assetsFile.Libraries.Where(l => l.Type == "project");
-            var projectLibraryPathToName = new Dictionary<string, string>(projectLibraries.Count());
-
-            foreach (var library in projectLibraries)
-            {
-                projectLibraryPathToName.Add(Path.GetFullPath(library.Path, projectDirectoryPath), library.Name);
-            }
-
-            foreach (string targetAlias in targetAliases)
-            {
-                topLevelReferences.Add(targetAlias, []);
-
-                // top level packages
-                TargetFrameworkInformation? targetFrameworkInformation = assetsFile.PackageSpec.TargetFrameworks.FirstOrDefault(tfi => tfi.TargetAlias.Equals(targetAlias, StringComparison.OrdinalIgnoreCase));
-                if (targetFrameworkInformation != default)
-                {
-                    var topLevelPackages = targetFrameworkInformation.Dependencies.Select(d => d.Name);
-                    topLevelReferences[targetAlias].AddRange(topLevelPackages);
-                }
-
-                // top level projects
-                ProjectRestoreMetadataFrameworkInfo? restoreMetadataFrameworkInfo = assetsFile.PackageSpec.RestoreMetadata.TargetFrameworks.FirstOrDefault(tfi => tfi.TargetAlias.Equals(targetAlias, StringComparison.OrdinalIgnoreCase));
-                if (restoreMetadataFrameworkInfo != default)
-                {
-                    var topLevelProjectPaths = restoreMetadataFrameworkInfo.ProjectReferences.Select(p => p.ProjectPath);
-                    foreach (var projectPath in topLevelProjectPaths)
-                    {
-                        topLevelReferences[targetAlias].Add(projectLibraryPathToName[projectPath]);
-                    }
-                }
-            }
-
-            return topLevelReferences;
-        }
-
-        private class StackDependencyData
-        {
-            public string Id { get; set; }
-            public StackDependencyData? Parent { get; set; }
-
-            public StackDependencyData(string currentId, StackDependencyData? parentDependencyData)
-            {
-                Id = currentId;
-                Parent = parentDependencyData;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Represents a node in the package dependency graph.
-    /// </summary>
-    internal class DependencyNode
-    {
-        public string Id { get; set; }
-        public string Version { get; set; }
-        public HashSet<DependencyNode> Children { get; set; }
-
-        public DependencyNode(string id, string version)
-        {
-            Id = id;
-            Version = version;
-            Children = new HashSet<DependencyNode>();
-        }
-
-        public override int GetHashCode()
-        {
-            var hashCodeCombiner = new HashCodeCombiner();
-            hashCodeCombiner.AddObject(Id);
-            hashCodeCombiner.AddObject(Version);
-            hashCodeCombiner.AddUnorderedSequence(Children);
-            return hashCodeCombiner.CombinedHash;
         }
     }
 }
